+++ src/date/addUTC.js
@@ -50,6 +50,11 @@ export default function addUTC (val, amount, key) {
             copy.setUTCSeconds(copy.getUTCSeconds() + amount);
             return copy;
         }
+        case 'milliseconds':
+        case 'millisecond': {
+            copy.setUTCMilliseconds(copy.getUTCMilliseconds() + amount);
+            return copy;
+        }
         default:
             return copy;
     }